+++ src/Loop/EventLoop.php
@@ -4,7 +4,7 @@
  * This file is part of Icicle, a library for writing asynchronous code in PHP using promises and coroutines.
  *
  * @copyright 2014-2015 Aaron Piotrowski. All rights reserved.
- * @license Apache-2.0 See the LICENSE file that was distributed with this source code for more information.
+ * @license MIT See the LICENSE file that was distributed with this source code for more information.
  */
 
 namespace Icicle\Loop;
@@ -13,7 +13,7 @@ use Event;
 use EventBase;
 use Icicle\Loop\Events\EventFactoryInterface;
 use Icicle\Loop\Exception\UnsupportedError;
-use Icicle\Loop\Manager\Event\{AwaitManager, PollManager, SignalManager, TimerManager};
+use Icicle\Loop\Manager\Event\{SignalManager, SocketManager, TimerManager};
 use Icicle\Loop\Manager\{SignalManagerInterface, SocketManagerInterface, TimerManagerInterface};
 
 /**
@@ -27,16 +27,6 @@ class EventLoop extends AbstractLoop
     private $base;
 
     /**
-     * Determines if the event extension is loaded, which is required for this class.
-     *
-     * @return bool
-     */
-    public static function enabled(): bool
-    {
-        return extension_loaded('event');
-    }
-    
-    /**
      * @param bool $enableSignals True to enable signal handling, false to disable.
      * @param \Icicle\Loop\Events\EventFactoryInterface|null $eventFactory
      * @param \EventBase|null $base Use null for an EventBase object to be automatically created.
@@ -49,7 +39,7 @@ class EventLoop extends AbstractLoop
         EventBase $base = null
     ) {
         // @codeCoverageIgnoreStart
-        if (!self::enabled()) {
+        if (!extension_loaded('event')) {
             throw new UnsupportedError(__CLASS__ . ' requires the event extension.');
         } // @codeCoverageIgnoreEnd
         
@@ -96,7 +86,7 @@ class EventLoop extends AbstractLoop
      */
     protected function createPollManager(EventFactoryInterface $factory): SocketManagerInterface
     {
-        return new PollManager($this, $factory);
+        return new SocketManager($this, $factory, Event::READ);
     }
     
     /**
@@ -104,7 +94,7 @@ class EventLoop extends AbstractLoop
      */
     protected function createAwaitManager(EventFactoryInterface $factory): SocketManagerInterface
     {
-        return new AwaitManager($this, $factory);
+        return new SocketManager($this, $factory, Event::WRITE);
     }
     
     /**