+++ src/Loop/Events/SignalInterface.php
@@ -4,18 +4,18 @@
  * This file is part of Icicle, a library for writing asynchronous code in PHP using promises and coroutines.
  *
  * @copyright 2014-2015 Aaron Piotrowski. All rights reserved.
- * @license Apache-2.0 See the LICENSE file that was distributed with this source code for more information.
+ * @license MIT See the LICENSE file that was distributed with this source code for more information.
  */
 
 namespace Icicle\Loop\Events;
 
-interface SignalInterface
+interface SignalInterface extends EventInterface
 {
     /**
      * Calls the callback associated with the timer.
      */
     public function call();
-    
+
     /**
      * Alias of call().
      */