+++ src/Loop/Events/Timer.php
@@ -4,7 +4,7 @@
  * This file is part of Icicle, a library for writing asynchronous code in PHP using promises and coroutines.
  *
  * @copyright 2014-2015 Aaron Piotrowski. All rights reserved.
- * @license Apache-2.0 See the LICENSE file that was distributed with this source code for more information.
+ * @license MIT See the LICENSE file that was distributed with this source code for more information.
  */
 
 namespace Icicle\Loop\Events;
@@ -45,6 +45,11 @@ class Timer implements TimerInterface
      * @var bool
      */
     private $periodic;
+
+    /**
+     * @var bool
+     */
+    private $referenced = true;
     
     /**
      * @param \Icicle\Loop\Manager\TimerManagerInterface $manager
@@ -105,6 +110,10 @@ class Timer implements TimerInterface
     public function start()
     {
         $this->manager->start($this);
+
+        if (!$this->referenced) {
+            $this->manager->unreference($this);
+        }
     }
 
     /**
@@ -120,6 +129,7 @@ class Timer implements TimerInterface
      */
     public function unreference()
     {
+        $this->referenced = false;
         $this->manager->unreference($this);
     }
     
@@ -128,6 +138,7 @@ class Timer implements TimerInterface
      */
     public function reference()
     {
+        $this->referenced = true;
         $this->manager->reference($this);
     }