+++ src/Loop/LibeventLoop.php
@@ -4,14 +4,14 @@
  * This file is part of Icicle, a library for writing asynchronous code in PHP using promises and coroutines.
  *
  * @copyright 2014-2015 Aaron Piotrowski. All rights reserved.
- * @license Apache-2.0 See the LICENSE file that was distributed with this source code for more information.
+ * @license MIT See the LICENSE file that was distributed with this source code for more information.
  */
 
 namespace Icicle\Loop;
 
 use Icicle\Loop\Events\EventFactoryInterface;
 use Icicle\Loop\Exception\UnsupportedError;
-use Icicle\Loop\Manager\Libevent\{AwaitManager, PollManager, SignalManager, TimerManager};
+use Icicle\Loop\Manager\Libevent\{SignalManager, SocketManager, TimerManager};
 use Icicle\Loop\Manager\{SignalManagerInterface, SocketManagerInterface, TimerManagerInterface};
 
 /**
@@ -27,16 +27,6 @@ class LibeventLoop extends AbstractLoop
     private $base;
 
     /**
-     * Determines if the libevent extension is loaded, which is required for this class.
-     *
-     * @return bool
-     */
-    public static function enabled(): bool
-    {
-        return extension_loaded('libevent');
-    }
-    
-    /**
      * @param bool $enableSignals True to enable signal handling, false to disable.
      * @param \Icicle\Loop\Events\EventFactoryInterface|null $eventFactory
      * @param resource|null Resource created by event_base_new() or null to automatically create an event base.
@@ -46,7 +36,7 @@ class LibeventLoop extends AbstractLoop
     public function __construct($enableSignals = true, EventFactoryInterface $eventFactory = null, $base = null)
     {
         // @codeCoverageIgnoreStart
-        if (!self::enabled()) {
+        if (!extension_loaded('libevent')) {
             throw new UnsupportedError(__CLASS__ . ' requires the libevent extension.');
         } // @codeCoverageIgnoreEnd
 
@@ -98,7 +88,7 @@ class LibeventLoop extends AbstractLoop
      */
     protected function createPollManager(EventFactoryInterface $factory): SocketManagerInterface
     {
-        return new PollManager($this, $factory);
+        return new SocketManager($this, $factory, EV_READ);
     }
     
     /**
@@ -106,7 +96,7 @@ class LibeventLoop extends AbstractLoop
      */
     protected function createAwaitManager(EventFactoryInterface $factory): SocketManagerInterface
     {
-        return new AwaitManager($this, $factory);
+        return new SocketManager($this, $factory, EV_WRITE);
     }
     
     /**