+++ src/Loop/Manager/Ev/SocketManager.php
@@ -6,7 +6,7 @@ use Icicle\Loop\EvLoop;
 use Icicle\Loop\Exception\{FreedError, ResourceBusyError};
 use Icicle\Loop\Manager\SocketManagerInterface;
 
-abstract class SocketManager implements SocketManagerInterface
+class SocketManager implements SocketManagerInterface
 {
     const MIN_TIMEOUT = 0.001;
 
@@ -31,6 +31,11 @@ abstract class SocketManager implements SocketManagerInterface
     private $timers = [];
 
     /**
+     * @var \Icicle\Loop\Events\SocketEventInterface[]
+     */
+    private $unreferenced = [];
+
+    /**
      * @var callable
      */
     private $socketCallback;
@@ -39,26 +44,22 @@ abstract class SocketManager implements SocketManagerInterface
      * @var callable
      */
     private $timerCallback;
+
+    /**
+     * @var int
+     */
+    private $type;
     
     /**
-     * Creates an Event object on the given EventBase for the SocketEventInterface.
-     *
-     * @param   \EvLoop $loop
-     * @param   \Icicle\Loop\Events\SocketEventInterface $event
-     * @param   callable $callback
-     *
-     * @return  \EvIO
+     * @param \Icicle\Loop\EvLoop $loop
+     * @param \Icicle\Loop\Events\EventFactoryInterface $factory
+     * @param int $eventType
      */
-    abstract protected function createEvent(\EvLoop $loop, SocketEventInterface $event, callable $callback): \EvIO;
-    
-    /**
-     * @param   \Icicle\Loop\EvLoop $loop
-     * @param   \Icicle\Loop\Events\EventFactoryInterface $factory
-     */
-    public function __construct(EvLoop $loop, EventFactoryInterface $factory)
+    public function __construct(EvLoop $loop, EventFactoryInterface $factory, int $eventType)
     {
         $this->factory = $factory;
         $this->loop = $loop->getEvLoop();
+        $this->type = $eventType;
         
         $this->socketCallback = function (\EvIO $event) {
             /** @var \Icicle\Loop\Events\SocketEventInterface $socket */
@@ -104,8 +105,8 @@ abstract class SocketManager implements SocketManagerInterface
      */
     public function isEmpty(): bool
     {
-        foreach ($this->events as $event) {
-            if ($event->is_active) {
+        foreach ($this->events as $id => $event) {
+            if ($event->is_active && !isset($this->unreferenced[$id])) {
                 return false;
             }
         }
@@ -126,7 +127,8 @@ abstract class SocketManager implements SocketManagerInterface
 
         $socket = $this->factory->socket($this, $resource, $callback);
 
-        $event = $this->createEvent($this->loop, $socket, $this->socketCallback);
+        $event = $this->loop->io($resource, $this->type, $this->socketCallback, $socket);
+        $event->stop();
 
         $this->events[$id] = $event;
 
@@ -147,7 +149,6 @@ abstract class SocketManager implements SocketManagerInterface
         $this->events[$id]->start();
 
         if (0 !== $timeout) {
-            $timeout = (float) $timeout;
             if (self::MIN_TIMEOUT > $timeout) {
                 $timeout = self::MIN_TIMEOUT;
             }
@@ -198,7 +199,7 @@ abstract class SocketManager implements SocketManagerInterface
         
         if (isset($this->events[$id]) && $socket === $this->events[$id]->data) {
             $this->events[$id]->stop();
-            unset($this->events[$id]);
+            unset($this->events[$id], $this->unreferenced[$id]);
 
             if (isset($this->timers[$id])) {
                 $this->timers[$id]->stop();
@@ -216,7 +217,27 @@ abstract class SocketManager implements SocketManagerInterface
         
         return !isset($this->events[$id]) || $socket !== $this->events[$id]->data;
     }
-    
+
+    /**
+     * {@inheritdoc}
+     */
+    public function reference(SocketEventInterface $socket)
+    {
+        unset($this->unreferenced[(int) $socket->getResource()]);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function unreference(SocketEventInterface $socket)
+    {
+        $id = (int) $socket->getResource();
+
+        if (isset($this->events[$id]) && $socket === $this->events[$id]->data) {
+            $this->unreferenced[$id] = $socket;
+        }
+    }
+
     /**
      * {@inheritdoc}
      */
@@ -231,6 +252,7 @@ abstract class SocketManager implements SocketManagerInterface
         }
         
         $this->events = [];
+        $this->unreferenced = [];
         $this->timers = [];
     }
 }