+++ src/Loop/Manager/Libevent/SocketManager.php
@@ -4,7 +4,7 @@
  * This file is part of Icicle, a library for writing asynchronous code in PHP using promises and coroutines.
  *
  * @copyright 2014-2015 Aaron Piotrowski. All rights reserved.
- * @license Apache-2.0 See the LICENSE file that was distributed with this source code for more information.
+ * @license MIT See the LICENSE file that was distributed with this source code for more information.
  */
 
 namespace Icicle\Loop\Manager\Libevent;
@@ -16,7 +16,7 @@ use Icicle\Loop\Exception\{FreedError, ResourceBusyError};
 use Icicle\Loop\LibeventLoop;
 use Icicle\Loop\Manager\SocketManagerInterface;
 
-abstract class SocketManager implements SocketManagerInterface
+class SocketManager implements SocketManagerInterface
 {
     const MIN_TIMEOUT = 0.001;
     const MICROSEC_PER_SEC = 1e6;
@@ -50,33 +50,33 @@ abstract class SocketManager implements SocketManagerInterface
      * @var bool[]
      */
     private $pending = [];
+
+    /**
+     * @var \Icicle\Loop\Events\SocketEventInterface[]
+     */
+    private $unreferenced = [];
     
     /**
      * @var callable
      */
     private $callback;
-    
+
     /**
-     * Creates an event resource on the given event base for the SocketEventInterface.
-     *
-     * @param resource $base Event base resource.
-     * @param \Icicle\Loop\Events\SocketEventInterface $event
-     * @param callable $callback
-     *
-     * @return resource Event resource.
+     * @var int
      */
-    abstract protected function createEvent($base, SocketEventInterface $event, callable $callback);
+    private $type;
     
     /**
      * @param \Icicle\Loop\LibeventLoop $loop
      * @param \Icicle\Loop\Events\EventFactoryInterface $factory
-     * @param resource $base
+     * @param int $eventType
      */
-    public function __construct(LibeventLoop $loop, EventFactoryInterface $factory)
+    public function __construct(LibeventLoop $loop, EventFactoryInterface $factory, int $eventType)
     {
         $this->loop = $loop;
         $this->factory = $factory;
         $this->base = $this->loop->getEventBase();
+        $this->type = $eventType;
         
         $this->callback = function ($resource, int $what, SocketEventInterface $socket) {
             $this->pending[(int) $resource] = false;
@@ -99,8 +99,8 @@ abstract class SocketManager implements SocketManagerInterface
      */
     public function isEmpty(): bool
     {
-        foreach ($this->pending as $pending) {
-            if ($pending) {
+        foreach ($this->pending as $id => $pending) {
+            if ($pending && !isset($this->unreferenced[$id])) {
                 return false;
             }
         }
@@ -137,7 +137,11 @@ abstract class SocketManager implements SocketManagerInterface
         }
         
         if (!isset($this->events[$id])) {
-            $this->events[$id] = $this->createEvent($this->base, $socket, $this->callback);
+            $event = event_new();
+            event_set($event, $socket->getResource(), $this->type, $this->callback, $socket);
+            event_base_set($event, $this->base);
+
+            $this->events[$id] = $event;
         }
 
         $this->pending[$id] = true;
@@ -147,7 +151,6 @@ abstract class SocketManager implements SocketManagerInterface
             return;
         }
 
-        $timeout = (float) $timeout;
         if (self::MIN_TIMEOUT > $timeout) {
             $timeout = self::MIN_TIMEOUT;
         }
@@ -188,8 +191,7 @@ abstract class SocketManager implements SocketManagerInterface
         $id = (int) $socket->getResource();
         
         if (isset($this->sockets[$id]) && $socket === $this->sockets[$id]) {
-            unset($this->sockets[$id]);
-            unset($this->pending[$id]);
+            unset($this->sockets[$id], $this->pending[$id], $this->unreferenced[$id]);
             
             if (isset($this->events[$id])) {
                 event_free($this->events[$id]);
@@ -206,6 +208,26 @@ abstract class SocketManager implements SocketManagerInterface
         $id = (int) $socket->getResource();
         
         return !isset($this->sockets[$id]) || $socket !== $this->sockets[$id];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function reference(SocketEventInterface $socket)
+    {
+        unset($this->unreferenced[(int) $socket->getResource()]);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function unreference(SocketEventInterface $socket)
+    {
+        $id = (int) $socket->getResource();
+
+        if (isset($this->events[$id]) && $socket === $this->sockets[$id]) {
+            $this->unreferenced[$id] = $socket;
+        }
     }
     
     /**