+++ src/Loop/SelectLoop.php
@@ -4,7 +4,7 @@
  * This file is part of Icicle, a library for writing asynchronous code in PHP using promises and coroutines.
  *
  * @copyright 2014-2015 Aaron Piotrowski. All rights reserved.
- * @license Apache-2.0 See the LICENSE file that was distributed with this source code for more information.
+ * @license MIT See the LICENSE file that was distributed with this source code for more information.
  */
 
 namespace Icicle\Loop;
@@ -47,16 +47,6 @@ class SelectLoop extends AbstractLoop
      * @var \Icicle\Loop\Events\TimerInterface|null
      */
     private $signalTimer;
-
-    /**
-     * Always returns true for this class, since this class only requires core PHP functions.
-     *
-     * @return bool
-     */
-    public static function enabled(): bool
-    {
-        return true;
-    }
 
     /**
      * {@inheritdoc}