+++ tests/Loop/EventLoopTest.php
@@ -4,7 +4,7 @@
  * This file is part of Icicle, a library for writing asynchronous code in PHP using promises and coroutines.
  *
  * @copyright 2014-2015 Aaron Piotrowski. All rights reserved.
- * @license Apache-2.0 See the LICENSE file that was distributed with this source code for more information.
+ * @license MIT See the LICENSE file that was distributed with this source code for more information.
  */
 
 namespace Icicle\Tests\Loop;
@@ -30,10 +30,5 @@ class EventLoopTest extends AbstractLoopTest
     public function createLoop(EventFactoryInterface $eventFactory)
     {
         return new EventLoop(true, $eventFactory, self::$base);
-    }
-    
-    public function testEnabled()
-    {
-        $this->assertSame(extension_loaded('event'), EventLoop::enabled());
     }
 }